+++ src/integeruser/jgltut/tut17/DoubleProjection.java
@@ -195,7 +195,7 @@ public class DoubleProjection extends Tutorial {
             glBindBuffer(GL_UNIFORM_BUFFER, 0);
         }
 
-        glViewport(0, 0, (int) displaySize.x, (int) displaySize.y);
+        glViewport(0, 0, displaySize.x, displaySize.y);
         scene.render(modelMatrix);
 
         if (drawCameraPos) {
@@ -239,7 +239,7 @@ public class DoubleProjection extends Tutorial {
             glDisable(GL_DEPTH_CLAMP);
         }
 
-        glViewport((int) displaySize.x + (displayWidth % 2), 0, (int) displaySize.x, (int) displaySize.y);
+        glViewport(displaySize.x + (displayWidth % 2), 0, displaySize.x, displaySize.y);
         scene.render(modelMatrix);
         glEnable(GL_DEPTH_CLAMP);
     }
